pyproject.toml
[project]
name = "trade"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "requests>=2.32.5",
    "fake-useragent>=2.2.0",
    "python-dotenv>=1.2.1",
]
